Move ThreadCpuTime to Java 7 only location
